requirements.txt
streamlit
opencv-python
numpy
ultralytics
gtts
playsound==1.2.2
openai-whisper
sounddevice
scipy
